refactor: 의미에 맞는 패키지명으로 옮김
