+++ src/main/java/baseball/domain/GameNumberRandomGenerator.java
@@ -1,8 +1,6 @@
-package baseball;
+package baseball.domain;
 
 import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Collections;
 import java.util.List;
 import java.util.Random;
