+++ src/main/java/baseball/domain/exception/InvalidGameInputException.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.domain.exception;
 
 public class InvalidGameInputException extends IllegalArgumentException {
 	public InvalidGameInputException(String message) {
