+++ src/test/java/baseball/domain/GameHintTest.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.domain;
 
 import static org.assertj.core.api.Assertions.*;
